pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "verdictcouncil"
version = "0.1.0"
description = "Multi-agent AI judicial decision-support system"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0,<2.0",
    "langchain-core>=1.0,<2.0",
    "langgraph>=1.0,<2.0",
    "langchain-openai>=1.0,<2.0",
    "langgraph-checkpoint-postgres>=3.0,<4.0",
    "langsmith>=0.3.0",
    "deepdiff>=8.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.10",
    "redis>=5.2.0",
    "pyjwt>=2.10.0",
    "httpx>=0.28.0",
    "openai~=2.32.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.7.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.18",
    "arq~=0.26.3",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "opentelemetry-exporter-otlp>=1.29.0",
    # Layer-2 prompt-injection classifier for RAG corpus sanitization at ingest.
    # Wraps protectai/deberta-v3-base-prompt-injection-v2 (Apache-2.0, 95%+ recall).
    # See docs/security/rag-corpus-sanitization.md for architecture notes and citations.
    "llm-guard==0.3.16",
]

[project.optional-dependencies]
dev = [
    "honcho>=2.0.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.14.0",
    "pip-audit>=2.7.0",
    "bandit>=1.8.0",
    "python-dotenv>=1.0.0",
    "httpx",  # for FastAPI test client
    "factory-boy>=3.3.0",
    "pypdf>=4.0.0",  # for PDF export tests
    # Property-based / fuzz testing
    "hypothesis>=6.130.0",
    # Load / performance testing
    "locust>=2.32.0",
    # Mutation testing
    "mutmut>=2.4.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]
ignore = [
    "UP042",  # str+Enum pattern needed for Pydantic/SQLAlchemy compatibility
    "B008",   # FastAPI uses function calls in defaults (Depends, Query)
]

[tool.ruff.lint.per-file-ignores]
"src/pipeline/graph/prompts.py" = ["E501"]  # prompt strings intentionally exceed line limit

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["passlib.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests requiring live infrastructure (LLM, Postgres, Redis) — skipped in CI",
]

[dependency-groups]
dev = [
    "langgraph-cli[inmem]>=0.4.24",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "pytest-mock>=3.15.1",
]

# ---------------------------------------------------------------------------
# Mutation testing (mutmut)
# ---------------------------------------------------------------------------
# Usage:
#   mutmut run --paths-to-mutate src/ --tests-dir tests/unit/
#   mutmut results
#   mutmut html
#
# Target kill score: ≥ 80% (surviving mutants indicate test blind spots).
# Run locally or in a dedicated slow CI job — full run takes ~30 min.
# ---------------------------------------------------------------------------
[tool.mutmut]
paths_to_mutate = "src/"
tests_dir = "tests/unit/"
runner = "python -m pytest -x -q --no-header --tb=no"
# Do not mutate pure auto-generated code or template strings
dict_synonyms = ""

# ---------------------------------------------------------------------------
# Hypothesis settings (property-based testing)
# ---------------------------------------------------------------------------
# Profile "ci" is activated by setting HYPOTHESIS_PROFILE=ci in the CI env.
# Uses fewer examples to keep CI fast; local runs use the default (200).
[tool.hypothesis]
# Default max_examples for all @given tests when no per-test @settings applied
max_examples = 200

